fixed integer division

+++ src/java/Model/CircleAreaService.java
@@ -6,7 +6,7 @@ package Model;
  */
 public class CircleAreaService {
     public double getArea(String r){
-        int rad=r!=null?Integer.parseInt(r):0;
+        double rad=r!=null?Double.parseDouble(r):0;
         
         return Math.PI*Math.pow(rad,2);
     }
